+++ CHANGELOG.md
@@ -40,6 +40,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
     by default in NixOS 22.11.
 * [home-manager/Git] Set `pull.rebase = "merges"` instead of the deprecated
     `pull.rebase = "preserve"`.
+* [home-manager/GPG] Set the `disable-ccid` option for `scdaemon`, so that it
+    properly works with `pcscd`.
 * Add missing `mkDefault`.
 
 ## [0.0.16] - 2022-12-27